+++ backend/app/models/TempPhrase/model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy import Column, BigInteger, String, DateTime, ForeignKey, Integer
 from ...database.config import Base
 from sqlalchemy.orm import relationship
 
@@ -6,7 +6,7 @@ class TempPhrase(Base):
     __tablename__ = "temp_phrase"
 
     id = Column(Integer, primary_key=True, index=True)
-    id_player = Column(Integer, ForeignKey("player.id"))
+    id_player = Column(BigInteger, ForeignKey("player.id"))
     phrase = Column(String(500))
     exp = Column(DateTime(timezone=True))
